pyproject.toml
[tool.poetry]
name = "ask_chat_ip_addresses"
version = "0.1.0"
description = ""
authors = ["guins2000 <salfjsdl@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openpyxl = "^3.0.10"
pandas = "^1.4.2"
ipinfo = "^4.2.1"
lh3api = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
